+++ packages/common/src/client/AbstractPowerSyncDatabase.ts
@@ -193,6 +193,24 @@ export abstract class AbstractPowerSyncDatabase extends BaseObserver<PowerSyncDB
     return this.connectionManager.syncStreamImplementation;
   }
 
+  /**
+   * The connector used to connect to the PowerSync service.
+   * 
+   * @returns The connector used to connect to the PowerSync service or null if `connect()` has not been called.
+   */
+  get connector() {
+    return this.connectionManager.connector;
+  }
+
+  /**
+   * The resolved connection options used to connect to the PowerSync service.
+   * 
+   * @returns The resolved connection options used to connect to the PowerSync service or null if `connect()` has not been called.
+   */
+  get connectionOptions() {
+    return this.connectionManager.connectionOptions;
+  }
+
   protected _schema: Schema;
 
   private _database: DBAdapter;
@@ -355,7 +373,10 @@ export abstract class AbstractPowerSyncDatabase extends BaseObserver<PowerSyncDB
     return this.waitForStatus(statusMatches, signal);
   }
 
-  private async waitForStatus(predicate: (status: SyncStatus) => any, signal?: AbortSignal): Promise<void> {
+  /**
+   * Waits for the first sync status for which the `status` callback returns a truthy value.
+   */
+  async waitForStatus(predicate: (status: SyncStatus) => any, signal?: AbortSignal): Promise<void> {
     if (predicate(this.currentStatus)) {
       return;
     }
@@ -364,16 +385,21 @@ export abstract class AbstractPowerSyncDatabase extends BaseObserver<PowerSyncDB
       const dispose = this.registerListener({
         statusChanged: (status) => {
           if (predicate(status)) {
-            dispose();
-            resolve();
+            abort();
           }
         }
       });
 
-      signal?.addEventListener('abort', () => {
+      function abort() {
         dispose();
         resolve();
-      });
+      }
+
+      if (signal?.aborted) {
+        abort();
+      } else {
+        signal?.addEventListener('abort', abort);
+      }
     });
   }