+++ packages/common/src/client/ConnectionManager.ts
@@ -118,6 +118,14 @@ export class ConnectionManager extends BaseObserver<ConnectionManagerListener> {
     this.syncDisposer = null;
   }
 
+  get connector() {
+    return this.pendingConnectionOptions?.connector ?? null;
+  }
+
+  get connectionOptions() {
+    return this.pendingConnectionOptions?.options ?? null;
+  }
+
   get logger() {
     return this.options.logger;
   }
@@ -324,14 +332,15 @@ export class ConnectionManager extends BaseObserver<ConnectionManagerListener> {
     };
   }
 
-  private get activeStreams() {
+  /**
+   * @internal exposed for testing
+   */
+  get activeStreams() {
     return [...this.locallyActiveSubscriptions.values()].map((a) => ({ name: a.name, params: a.parameters }));
   }
 
   private subscriptionsMayHaveChanged() {
-    if (this.syncStreamImplementation) {
-      this.syncStreamImplementation.updateSubscriptions(this.activeStreams);
-    }
+    this.syncStreamImplementation?.updateSubscriptions(this.activeStreams);
   }
 }