+++ packages/react/src/index.ts
@@ -5,6 +5,7 @@ export * from './hooks/PowerSyncContext.js';
 export { SuspenseQueryResult } from './hooks/suspense/SuspenseQueryResult.js';
 export { useSuspenseQuery } from './hooks/suspense/useSuspenseQuery.js';
 export { useWatchedQuerySuspenseSubscription } from './hooks/suspense/useWatchedQuerySuspenseSubscription.js';
+export { useSyncStream, UseSyncStreamOptions } from './hooks/streams.js';
 export { useStatus } from './hooks/useStatus.js';
 export { useQuery } from './hooks/watched/useQuery.js';
 export { useWatchedQuerySubscription } from './hooks/watched/useWatchedQuerySubscription.js';