+++ packages/adapter-sql-js/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @powersync/adapter-sql-js
 
+## 0.0.8
+
+### Patch Changes
+
+- Updated dependencies [c2bc2c1]
+  - @powersync/common@1.40.0
+
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [eff8cbf]
+  - @powersync/common@1.39.0
+
 ## 0.0.6
 
 ### Patch Changes