+++ packages/common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @powersync/common
 
+## 1.40.0
+
+### Minor Changes
+
+- c2bc2c1: Exposed the connection and connection option used by `connect()`
+
+## 1.39.0
+
+### Minor Changes
+
+- eff8cbf: Add alpha support for sync streams, allowing different sets of data to be synced dynamically.
+
 ## 1.38.1
 
 ### Patch Changes