+++ packages/powersync-op-sqlite/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @powersync/op-sqlite
 
+## 0.7.11
+
+### Patch Changes
+
+- Updated dependencies [c2bc2c1]
+  - @powersync/common@1.40.0
+
+## 0.7.10
+
+### Patch Changes
+
+- c78071f: Update core extension to 0.4.6
+- Updated dependencies [eff8cbf]
+  - @powersync/common@1.39.0
+
 ## 0.7.9
 
 ### Patch Changes