+++ packages/tanstack-react-query/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @powersync/tanstack-react-query
 
+## 0.1.6
+
+### Patch Changes
+
+- Updated dependencies [3008dbc]
+- Updated dependencies [c2bc2c1]
+  - @powersync/react@1.8.1
+  - @powersync/common@1.40.0
+
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies [eff8cbf]
+- Updated dependencies [d8236aa]
+- Updated dependencies [7a5aaf5]
+  - @powersync/common@1.39.0
+  - @powersync/react@1.8.0
+
 ## 0.1.4
 
 ### Patch Changes